+++ src/server/__tests__/test-cleaner.js
@@ -1,0 +1,6 @@
+describe('server', () => {
+    describe('cleaner', () => {
+        jest.unmock('../cleaner');
+        require('../cleaner');
+    });
+});