+++ src/server/index.js
@@ -1,2 +1,6 @@
-require('source-map-support').install();
-require('./io');
+import config from 'config';
+import { runAll as runCleaner } from './cleaner';
+
+import './io';
+
+runCleaner(config.get('data_cleaner'));